refactor: move clientBuild
